wl-sim: full refactor, add bash script for running
