+++ wl-sim/main/CMakeLists.txt
@@ -1,4 +1,4 @@
-set(srcs "main.cpp" "random.cpp" "WL_Flash.cpp")
+set(srcs "main.cpp" "wl_sim_random.cpp" "WLsim_Flash.cpp")
 
 idf_component_register(SRCS ${srcs}
                         INCLUDE_DIRS "include")
